+++ src/primitives/arrow.js
@@ -6,17 +6,19 @@ import {
 	Line,
 	Matrix4,
 	Mesh,
-	RawShaderMaterial,
 	Vector3
 } from '../../vendors/three.js';
 
 import { getPopulatedCoordinateBuffer } from '../bufferUtils.js';
 
 import scaleCoordinate from '../scaleCoordinate.js';
+import { getBasicMaterial } from '../shader.js';
 
 // See https://reference.wolfram.com/language/ref/Arrow
 // for the high-level description of what is being rendered.
 export default function ({ color, coords, opacity = 1 }, extent) {
+	const material = getBasicMaterial(color, opacity);
+
 	const group = new Group();
 
 	// last coordinate but one
@@ -52,31 +54,7 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 						new Vector3(0, 1, 0)
 					)
 			),
-		new RawShaderMaterial({
-			transparent: opacity !== 1,
-			uniforms: {
-				color: { value: [...color, opacity] }
-			},
-			vertexShader: `#version 300 es
-				in vec3 position;
-
-				uniform mat4 projectionMatrix;
-				uniform mat4 modelViewMatrix;
-
-				void main() {
-					gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1);
-				}
-			`,
-			fragmentShader: `#version 300 es
-				uniform lowp vec4 color;
-
-				out lowp vec4 pc_fragColor;
-
-				void main() {
-					pc_fragColor = color;
-				}
-			`
-		})
+		material
 	));
 
 	// arrow body
@@ -89,32 +67,7 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 					3
 				)
 			),
-			new RawShaderMaterial({
-				opacity,
-				transparent: opacity !== 1,
-				uniforms: {
-					color: { value: [...color, opacity] }
-				},
-				vertexShader: `#version 300 es
-					in vec3 position;
-
-					uniform mat4 projectionMatrix;
-					uniform mat4 modelViewMatrix;
-
-					void main() {
-						gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1);
-					}
-				`,
-				fragmentShader: `#version 300 es
-					uniform lowp vec4 color;
-
-					out lowp vec4 pc_fragColor;
-	
-					void main() {
-						pc_fragColor = color;
-					}
-				`
-			})
+			material
 		)
 	);