+++ src/primitives/line.js
@@ -1,11 +1,11 @@
 import {
 	BufferAttribute,
 	BufferGeometry,
-	Line,
-	RawShaderMaterial
+	Line
 } from '../../vendors/three.js';
 
 import { getPopulatedCoordinateBuffer } from '../bufferUtils.js';
+import { getBasicMaterial } from '../shader.js';
 
 // See https://reference.wolfram.com/language/ref/Line
 // for the high-level description of what is being rendered.
@@ -20,31 +20,6 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 				3
 			)
 		),
-		new RawShaderMaterial({
-			opacity,
-			transparent: opacity !== 1,
-			uniforms: {
-				color: { value: [...color, opacity] }
-			},
-			vertexShader: `#version 300 es
-				in vec3 position;
-
-				uniform mat4 projectionMatrix;
-				uniform mat4 modelViewMatrix;
-
-				void main() {
-					gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1);
-				}
-			`,
-			fragmentShader: `#version 300 es
-				uniform lowp vec4 color;
-
-				out lowp vec4 pc_fragColor;
-
-				void main() {
-					pc_fragColor = color;
-				}
-			`
-		})
+		getBasicMaterial(color, opacity)
 	);
 }